+++ src/main/java/com/exasol/edmlgenerator/parquet/ToUpperSnakeCaseConverter.java
@@ -3,7 +3,7 @@ package com.exasol.edmlgenerator.parquet;
 /**
  * This class converts strings to UPPER_SNAKE_CASE.
  */
-class ToUpperSnakeCaseConverter {
+public class ToUpperSnakeCaseConverter {
     private ToUpperSnakeCaseConverter() {
         // empty on purpose
     }
@@ -14,7 +14,7 @@ class ToUpperSnakeCaseConverter {
      * @param inputString input string in camelCase, lower_snake_case or UPPER_SNAKE_CASE.
      * @return string formatted in UPPER_SNAKE_CASE
      */
-    static String toUpperSnakeCase(final String inputString) {
+    public static String toUpperSnakeCase(final String inputString) {
         boolean isPreviousUpperOrUnderscore = false;
         boolean isFirst = true;
         final StringBuilder result = new StringBuilder();
